requirements.test.txt
# aioresponses
# codecov
# coverage>=5.2.0,<5.3.0
# mypy

# fuzzywuzzy
# levenshtein
# ruff
# yamllint

pytest
pytest-asyncio
pytest-cov
pytest-homeassistant-custom-component
